requirements.txt
joblib==1.5.3
matplotlib==3.10.8
numpy==2.4.3
pandas==2.3.3
plotly==6.6.0
pyserial==3.5
scikit-learn==1.8.0
shap==0.51.0
streamlit==1.55.0